add the test

+++ tests/test.cc
@@ -3,7 +3,7 @@
 #include <cmath>
 
 TEST(PowerTest, PositiveIntNos) {
-	for (int n = 0; n < 10; n++)
+	for (long long  n = 0; n < 10; n++)
 		for (int base = 0; base < 20; ++base)
 			EXPECT_DOUBLE_EQ(getPower(n, base), std::pow(n, base));
 }
@@ -15,7 +15,7 @@ TEST(PowerTest, PositiveFltNos) {
 }
 
 TEST(PowerTest, NegativeNos) {
-	EXPECT_EQ(getPower(1,-1), 0);
+	EXPECT_EQ(getPower(1ll,-1), 0);
 	EXPECT_EQ(getPower(1.0,-1.3), 0);
 }
 int main(int argc, char **argv)
